+++ lib/cocoapods/command/search.rb
@@ -35,6 +35,10 @@ module Pod
       def validate!
         super
         help! "A search query is required." unless @query
+
+        /#{@query.join(' ').strip}/
+      rescue RegexpError
+        help! "A valid regular expression is required."
       end
 
       def run