+++ lib/cocoapods/gem_version.rb
@@ -1,6 +1,6 @@
 module Pod
   # The version of the cocoapods command line tool.
   #
-  VERSION = '0.29.0' unless defined? Pod::VERSION
+  VERSION = '0.30.0' unless defined? Pod::VERSION
 end